Primarily added a secondary attack function.
Copied various methods and variables in MainPlayer and WeaponBase, renamed and re-used them for "Special Attacks" so that the Game Designers can now set different attack stats depending on if it's the primary attack or special attack.
Difficult to explain, but please look at the altered codes and test play it for yourselves. The special attack input is the thumb mouse button (inner side button on your mouse).

The special attack addition now makes the weapon system feel a lot more similar to Immortal's. =)

(note: also deleted commented out code in BeginPlay in Mainplayer)